refactor(ws): move ws classes to same package
